pyproject.toml
[project]
name = "microservice-zoho-desk-ticket-modifications"
version = "0.3.0"
description = "FastAPI service to update Zoho Desk tickets and add templated private comments"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    "pydantic>=2,<3",
    "pydantic-settings>=2.2.0",
    "pyyaml>=6.0",
]
